apps/web/src/features/website-auditing/hooks/use-website-auditing.tsx
interface UseWebsiteAuditingOptions {
  /** Optional configuration for the website-auditing hook */
  config?: {
    /** Enable/disable feature */
    enabled?: boolean
    /** Timeout in milliseconds */
    timeout?: number
  }
}

interface UseWebsiteAuditingResult {
  /** Loading state */
  isLoading: boolean
  /** Error state */
  error: Error | null
  /** Reset the hook state */
  reset: () => void
}

export const useWebsiteAuditing = ({
  config,
}: UseWebsiteAuditingOptions = {}): UseWebsiteAuditingResult => {
  return {
    isLoading: false,
    error: null,
    reset: () => {
      // Reset implementation
    },
  }
}
